Deduplicate SMTP and PDF helpers into emailbot modules
