+++ emailbot/settings.py
@@ -13,6 +13,10 @@ STRICT_OBFUSCATION: bool = True
 FOOTNOTE_RADIUS_PAGES: int = 1
 PDF_LAYOUT_AWARE: bool = False
 ENABLE_OCR: bool = False
+ENABLE_PROVIDER_CANON: bool = os.getenv("ENABLE_PROVIDER_CANON", "1") == "1"
+CANON_GMAIL_DOTS: bool = os.getenv("CANON_GMAIL_DOTS", "1") == "1"
+CANON_GMAIL_PLUS: bool = os.getenv("CANON_GMAIL_PLUS", "1") == "1"
+CANON_OTHER_PLUS: bool = os.getenv("CANON_OTHER_PLUS", "0") == "1"
 MAX_ASSETS: int = 8
 MAX_SITEMAP_URLS: int = 200
 MAX_DOCS: int = 30
@@ -43,12 +47,19 @@ def load() -> None:
     """Load configuration from the persistent store."""
 
     global STRICT_OBFUSCATION, FOOTNOTE_RADIUS_PAGES, PDF_LAYOUT_AWARE, ENABLE_OCR
+    global ENABLE_PROVIDER_CANON, CANON_GMAIL_DOTS, CANON_GMAIL_PLUS, CANON_OTHER_PLUS
     global MAX_ASSETS, MAX_SITEMAP_URLS, MAX_DOCS, PER_REQUEST_TIMEOUT
     global EXTERNAL_SOURCES, DAILY_SEND_LIMIT, SKIPPED_PREVIEW_LIMIT, LAST_SUMMARY_DIR
     STRICT_OBFUSCATION = bool(_store.get("STRICT_OBFUSCATION", STRICT_OBFUSCATION))
     FOOTNOTE_RADIUS_PAGES = int(_store.get("FOOTNOTE_RADIUS_PAGES", FOOTNOTE_RADIUS_PAGES))
     PDF_LAYOUT_AWARE = bool(_store.get("PDF_LAYOUT_AWARE", PDF_LAYOUT_AWARE))
     ENABLE_OCR = bool(_store.get("ENABLE_OCR", ENABLE_OCR))
+    ENABLE_PROVIDER_CANON = bool(
+        _store.get("ENABLE_PROVIDER_CANON", ENABLE_PROVIDER_CANON)
+    )
+    CANON_GMAIL_DOTS = bool(_store.get("CANON_GMAIL_DOTS", CANON_GMAIL_DOTS))
+    CANON_GMAIL_PLUS = bool(_store.get("CANON_GMAIL_PLUS", CANON_GMAIL_PLUS))
+    CANON_OTHER_PLUS = bool(_store.get("CANON_OTHER_PLUS", CANON_OTHER_PLUS))
     MAX_ASSETS = int(_store.get("MAX_ASSETS", MAX_ASSETS))
     MAX_SITEMAP_URLS = int(_store.get("MAX_SITEMAP_URLS", MAX_SITEMAP_URLS))
     MAX_DOCS = int(_store.get("MAX_DOCS", MAX_DOCS))
@@ -68,6 +79,10 @@ def save() -> None:
     _store.set("FOOTNOTE_RADIUS_PAGES", FOOTNOTE_RADIUS_PAGES)
     _store.set("PDF_LAYOUT_AWARE", PDF_LAYOUT_AWARE)
     _store.set("ENABLE_OCR", ENABLE_OCR)
+    _store.set("ENABLE_PROVIDER_CANON", ENABLE_PROVIDER_CANON)
+    _store.set("CANON_GMAIL_DOTS", CANON_GMAIL_DOTS)
+    _store.set("CANON_GMAIL_PLUS", CANON_GMAIL_PLUS)
+    _store.set("CANON_OTHER_PLUS", CANON_OTHER_PLUS)
     _store.set("MAX_ASSETS", MAX_ASSETS)
     _store.set("MAX_SITEMAP_URLS", MAX_SITEMAP_URLS)
     _store.set("MAX_DOCS", MAX_DOCS)
@@ -117,6 +132,10 @@ __all__ = [
     "FOOTNOTE_RADIUS_PAGES",
     "PDF_LAYOUT_AWARE",
     "ENABLE_OCR",
+    "ENABLE_PROVIDER_CANON",
+    "CANON_GMAIL_DOTS",
+    "CANON_GMAIL_PLUS",
+    "CANON_OTHER_PLUS",
     "MAX_ASSETS",
     "MAX_SITEMAP_URLS",
     "MAX_DOCS",
